refactor: schedule 관련 서버 상태 관리 로직 폴더 분리
